// src/app/models/encounter/ORMGYNForm.ts
export class ORMGYNForm {
    id: String;
    patient_id: String;
    chart_id: String;
    form_id: String;
    form_date: String;
    practice_id: String;
    deleted: Boolean;
    created_user: String;
    client_date_created: String;
    modified_user: String;
    client_date_modified: String;
    date_created: String;
    date_modified: String;
    system_ip: String;
    file_link: String;
    pregnancy_id: String;
}